Organized files and added noise filtering

+++ C++/noise_filter.cpp
@@ -2,21 +2,6 @@
 #include <stdint.h>
 #include <windows.h>
 
-// std::vector<float> filter_signal(std::vector<float> raw_signal){
-//     float a0 = 0.8948577513857248;
-//     float a1 = -1.7897155027714495;
-//     float a2 = 0.8948577513857248;
-//     float b1 = -1.7786300789392977;
-//     float b2 = 0.8008009266036016;
-
-//     std::vector<float> filtered_signal;
-
-//     for (int i =0; i< raw_signal.size(); i++){
-//         filtered_signal.push_back(a0*raw_signal[i] + a1*raw_signal[i-1] + a2*raw_signal[i-2] - b1*filtered_signal[i-1] - b2*filtered_signal[i-2]);
-//     }
-//     return filtered_signal;
-// }
-
 // function to filter the noise out of the signal 
 // INPUT:   pointer to the raw signal
 //          size of the raw signal
@@ -53,8 +38,8 @@ float* filter_signal(float* raw_signal, int size){
 // new filtered signal
     float* filtered_signal = (float*)malloc(sizeof(float)*size);
     // Initial values set to 0
-    filtered_signal[0] = (uint16_t)0;
-    filtered_signal[0] = (uint16_t)0;
+    filtered_signal[0] = (float)0;
+    filtered_signal[0] = (float)0;
 
    for (int i =2; i< size; i++){
         filtered_signal[i] = a0*raw_signal[i] + a1*raw_signal[i-1] + a2*raw_signal[i-2] - b1*filtered_signal[i-1] - b2*filtered_signal[i-2];
